refactor: move add to cart buttons actions to separate hook
