+++ src/components/add-to-cart-button/styles.module.css
@@ -1,0 +1,8 @@
+.add-to-cart-button {
+  margin-top: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  width: fit-content;
+}
