+++ src/pages/product-page/styles.module.css
@@ -140,8 +140,3 @@
   border-color: transparent;
   background-color: var(--g-color-base-brand);
 }
-
-.add-to-cart-button {
-  margin-top: 20px;
-  width: fit-content;
-}
